Fixing newline on Windows
Simple version: Newlines now also work with Windows encoded files in
Windows command prompt on Windows.

Complex version: If on an OS with \r\n set as a newline, after
processing the character entered the program will also check if the next
byte in the file is \n, and if so process that as well, with it checking
the character entered against \r if in memorization mode.